Pull main

+++ frontend/pages/archivos/verArchivos.jsx
@@ -4,8 +4,43 @@ import axios from 'axios'
 import { useRouter } from 'next/router'
 import Arriba from '../../components/Arriba'
 import Swal from 'sweetalert2'
+import {checkToken} from '../../data/usuario'
+const jwt = require('jwt-simple')
+import { DownloadIcon, DeleteIcon} from '../../node_modules/@chakra-ui/icons'
 
-const archivos= () => {
+
+//funcion para obtener la cookie y ver que rol tiene el usuario
+export async function getServerSideProps(context){
+    try {
+        const res = await checkToken(context.req.headers.cookie)
+        const decode = jwt.decode(context.req.cookies.token,process.env.SECRET_KEY)
+            if (res.status === 200){
+                return{
+                    props:{
+                        existe: res.config.headers.cookie,
+                        rol:decode.rol
+                    }
+                }
+            }else{
+                return{
+                    redirect: {
+                        destination: "/",
+                        permanent: false
+                    }
+                }
+            }
+    } catch (error) {
+        console.log("ERROR",error)
+        return{
+            redirect:{
+                destination: '/',
+                permanent: true
+            }
+        }
+    }
+}
+
+const archivos= (data) => {
 
     const [archivos, setArchivos] = useState([])
     const router = useRouter()
@@ -25,28 +60,31 @@ const archivos= () => {
     }
 
     const eliminarArchivos = async (id_archivo, id_asamblea) =>{
-        try {
-            //AGREGAR MODALs
-            const response = await axios.delete(`${process.env.API_URL}/file/delete/${id_archivo}/${id_asamblea}`)
-            if (response.status === 200){
+        //Si el usuario es admin podra eliminar archivos
+        if(data.rol==='admin'){
+            try {
+                //AGREGAR MODALs
+                const response = await axios.delete(`${process.env.API_URL}/file/delete/${id_archivo}/${id_asamblea}`)
+                if (response.status === 200){
+                    Swal.fire({
+                        title: 'Archivo eliminado',
+                        text: 'El archivo se ha eliminado con exito',
+                        icon: 'success',
+                        confirmButtonText: 'Ok'
+                    }).then((result)=>{
+                        if (result.isConfirmed){
+                            router.reload()
+                        }
+                    })
+                }
+                }catch (error) {
                 Swal.fire({
-                    title: 'Archivo eliminado',
-                    text: 'El archivo se ha eliminado con exito',
-                    icon: 'success',
+                    title: 'Error',
+                    text: `El archivo no se ha podido eliminar`,
+                    icon: 'error',
                     confirmButtonText: 'Ok'
-                }).then((result)=>{
-                    if (result.isConfirmed){
-                        router.reload()
-                    }
                 })
             }
-            }catch (error) {
-            Swal.fire({
-                title: 'Error',
-                text: `El archivo no se ha podido eliminar`,
-                icon: 'error',
-                confirmButtonText: 'Ok'
-            })
         }
     }
 
@@ -66,9 +104,15 @@ const archivos= () => {
                     </Td>
                     <Td>{archivos.fecha}</Td>
                     <Td>
-                        <Button onClick={()=>router.push(`${process.env.API_URL}/file/download/${archivos._id}`)}>Download</Button>
+                        <Button mx={'2.5'} bg={'white'} onClick={()=>router.push(`${process.env.API_URL}/file/download/${archivos._id}`)}>
+                            <DownloadIcon  w={6} h={6} color="green.500" ></DownloadIcon>
+                        </Button>
                     </Td>
-                    <Td><Button onClick={()=>eliminarArchivos(archivos._id, archivos.asamblea._id)}>Eliminar</Button></Td>
+                    <Td>
+                        <Button mx={'2.5'}  bg={'white'} onClick={()=>eliminarArchivos(archivos._id, archivos.asamblea._id)}>
+                            <DeleteIcon  w={6} h={6} color="red.400"></DeleteIcon>
+                        </Button>
+                    </Td>
                 </Tr>
             )
         })
@@ -76,7 +120,7 @@ const archivos= () => {
 
     return (
         <Box>
-            <Arriba/>
+            <Arriba token={data.existe}/>
             <Container maxW="container.xl">
             <Heading textAlign={"center"} my={16}>Historial de Actas</Heading>
             <Button colorScheme={"teal"} my={15} mx={15} float={'left'} onClick={()=>router.push('/')} >Volver</Button>
@@ -86,7 +130,7 @@ const archivos= () => {
                         <Td>Archivo</Td>
                         <Td>Asamblea</Td>
                         <Td>Fecha</Td>
-                        <Td>Descarga</Td>
+                        <Td>Descargar</Td>
                         <Td>Eliminar</Td>
                     </Tr>
                 </Thead>
